Refactor the HammerArea types and functions into multiple files
